Fix cancelling single file uploads
Unfortunately the ECMAScript websocket API provides no way to cancel a
buffered datastream. The close() function will actually wait until the
socket's buffer is empty before acting. This means our options are
limited to sending small chunks interrupted with confirmation messages
from the server, a CPU intensive interval that feeds the socket data as
the buffer nears empty, or reworking things to use HTTP file uploads.
This implements the first option as that seems to introduce the least
amount of added complexity. But mostly this is just down to the
websocket API missing a very basic feature.