Change snowflake IDs back to string type
The new version of discord.py changes all snowflake IDs to be integer
typed. Unfortunately this doesn't work for the project as this change
conflicts with both the official Discord API and the websocket messaging
API for the frontend. As well the JavaScript frontend doesn't support
64-bit integers. So this requires either additional packaging bloat for
the frontend or lots of type conversions on the backend. Tried to keep
the type conversions isolated to where discord.py interfaces with the
rest of the app, but there are a few places where discord.py usage has
leaked outside of where it should be. All in all a really ugly change
with no good solution because of incredibly questionable API changes in
a third party library.