Add trial mode setting
This allows me to host a public version of the bot that people can try
before going to the effort of installing it. Trial mode is disabled by
default for fresh installs. When trial mode is enabled the bot will
automatically leave any server it joins (unless whitelisted in
config.json) after 1 hour.