Fix passing uninitialized opus encoder
Looking through the blame of discord.py I'm not sure why this ever
worked before since this is normally initialized after playback starts
but we need it before then. Feels dirty using a separate instance for
our stream and the voice client, but it's only used to read a few static
variables.